extract common methods into mixin
