+++ spec/lib/audit/catalog_to_moab_instance_spec.rb
@@ -66,7 +66,7 @@ RSpec.describe CatalogToMoab do
     end
 
     it 'calls online_moab_found?(druid, storage_dir)' do
-      expect(c2m).to receive(:online_moab_found?).with(druid, storage_dir)
+      expect(c2m).to receive(:online_moab_found?)
       c2m.check_catalog_version
     end
